refactor: seperate interval jobs from main utils files (#468)
* chore: install bree

* refactor: move purgeusernames and topglobal to jobs folder

* refactor: put it in right format thing to work

* refactor: rename webhook url to better name

* refactor: topamountglobal no longer requires client

* refactor: topglobal to work with bree

* feat: setup bree to run jobs

* feat: start jobs once nypsi starts

* chore: install @types/node

* refactor: use parentport in place of logger

* feat: workers job

* fix: topglobal works now

* refactor: use redis for cache instead of set

* refactor: idk):

* fix: oldguilds work with scheduler

* feat: move deterioratekarma to jobs

* refactor: moved lottery job to clusterjobs folder

* refactor: move guildcountdowns to clusterjobs folder

* refactor: dont log when cant find guild

* refactor: move christmas thing to clusterjobs folder

* refactor: move chatreactions interval to clusterjobs folder

* refactor: moved moderationchecks to clusterjobs folder

* refactor: convert lottery ticket queue to bree job
