+++ package.json
@@ -25,6 +25,7 @@
     "@types/lodash": "^4.14.182",
     "@types/node-fetch": "^2.6.2",
     "better-sqlite3": "^7.6.2",
+    "bree": "^9.1.1",
     "bufferutil": "^4.0.6",
     "chalk": "^4.1.2",
     "crypto-js": "^4.1.1",
@@ -57,6 +58,7 @@
     "@types/crypto-js": "^4.1.1",
     "@types/figlet": "^1.5.4",
     "@types/ms": "^0.7.31",
+    "@types/node": "^18.0.6",
     "@types/node-os-utils": "^1.2.0",
     "@types/shuffle-array": "^1.0.2",
     "@types/urban-dictionary": "^3.0.0",
