+++ src/utils/karma/utils.ts
@@ -1,9 +1,6 @@
 import { GuildMember } from "discord.js";
-import ms = require("ms");
 import prisma from "../database/database";
 import redis from "../database/redis";
-import { MStoTime } from "../functions/date";
-import { logger } from "../logger";
 import { createProfile } from "../users/utils";
 
 let karmaShop = false;
@@ -141,87 +138,3 @@ export async function getLastCommand(member: GuildMember | string): Promise<Date
 
     return query.lastCommand;
 }
-
-async function deteriorateKarma() {
-    const now = Date.now();
-
-    const threshold = now - ms("7 hours");
-
-    const users = await prisma.user.findMany({
-        where: {
-            karma: { gt: 1 },
-        },
-        select: {
-            id: true,
-            karma: true,
-            lastCommand: true,
-        },
-    });
-
-    let total = 0;
-
-    for (const user of users) {
-        if (user.lastCommand.getTime() > threshold) continue;
-
-        let karmaToRemove = 5;
-
-        if (now - ms("1 week") > user.lastCommand.getTime()) {
-            karmaToRemove = 35;
-        }
-
-        if (now - ms("30 days") > user.lastCommand.getTime()) {
-            karmaToRemove = 100;
-        }
-
-        if (now - ms("90 days") > user.lastCommand.getTime()) {
-            karmaToRemove = 69420;
-        }
-
-        if (karmaToRemove > user.karma) {
-            karmaToRemove = user.karma - 1;
-        }
-
-        total += karmaToRemove;
-
-        await prisma.user.update({
-            where: {
-                id: user.id,
-            },
-            data: {
-                karma: { decrement: karmaToRemove },
-            },
-        });
-
-        await redis.del(`cache:user:karma:${user.id}`);
-    }
-
-    logger.log({
-        level: "auto",
-        message: `${total} total karma deteriorated`,
-    });
-}
-
-// prettier-ignore
-(() => {
-    const now = new Date();
-
-    let d = `${now.getMonth() + 1}/${now.getDate() + 1}/${now.getUTCFullYear()}`;
-
-    if (now.getHours() < 3) {
-        d = `${now.getMonth() + 1}/${now.getDate()}/${now.getUTCFullYear()}`;
-    }
-
-    const needed = new Date(Date.parse(d) + 10800000).getTime();
-
-    setTimeout(async () => {
-        setInterval(() => {
-            deteriorateKarma();
-        }, 86400000);
-        deteriorateKarma();
-    }, needed - now.getTime());
-
-    logger.log({
-        level: "auto",
-        message: `karma deterioration will run in ${MStoTime(needed - now.getTime())}`
-    });
-})();
